% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{add_deployment_columns}
\alias{add_deployment_columns}
\title{Add deployment date and sensor columns and re-order all columns}
\usage{
add_deployment_columns(dat, start_date, end_date, sn_table)
}
\arguments{
\item{dat}{placeholder}

\item{start_date}{placeholder}

\item{end_date}{placeholder}

\item{sn_table}{A data frame with three columns: \code{sensor_type},
code{sensor_serial_number}, \code{depth}.}
}
\value{
returns dat with additional columns
}
\description{
Add deployment date and sensor columns and re-order all columns
}
