% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{check_n_rows}
\alias{check_n_rows}
\title{Check number of rows of data file}
\usage{
check_n_rows(dat, file_name, trimmed = TRUE)
}
\arguments{
\item{dat}{Data frame}

\item{file_name}{Name of file to check.}

\item{trimmed}{Logical value indicating if \code{dat} has been trimmed.}
}
\value{
Returns a Warning if there no rows in \code{dat}.
}
\description{
Check number of rows of data file
}
