% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-plot.R
\name{create_variable_labels}
\alias{create_variable_labels}
\title{Create plot labels from variable names}
\usage{
create_variable_labels(dat, new_line = FALSE)
}
\arguments{
\item{dat}{Data frame of Water Quality data with variables in long format.}

\item{new_line}{Logical argument indicating whether to place the units of the
y-axis on a new line. If \code{FALSE}, the default, the variable name and
units are on the same line.}
}
\value{
Returns \code{dat} with an additional column \code{variable_label}.
}
\description{
Create plot labels from variable names
}
