% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_aquameasure_tz}
\alias{extract_aquameasure_tz}
\title{Extract the timezone of aquameasure timestamps}
\usage{
extract_aquameasure_tz(am_colnames)
}
\arguments{
\item{am_colnames}{Column names of aquameasure data file.}
}
\value{
Returns a character string of the timezone indicated in the Timestamp
  column.
}
\description{
Extract the timezone of aquameasure timestamps
}
