% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{extract_aquameasure_vars}
\alias{extract_aquameasure_vars}
\title{Extract the variables included in aquameasure file from the column names}
\usage{
extract_aquameasure_vars(am_colnames)
}
\arguments{
\item{am_colnames}{Column names of aquameasure data file.}
}
\value{
Returns a vector of the variables included in the file.
}
\description{
Extract the variables included in aquameasure file from the column names
}
