% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{extract_deployment_dates}
\alias{extract_deployment_dates}
\title{Extract deployment dates}
\usage{
extract_deployment_dates(deployment_dates)
}
\arguments{
\item{deployment_dates}{Data frame with two columns; the first column holds
the deployment date, and the second column holds the retrieval date.}
}
\description{
Extract deployment dates
}
