% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-misc.R
\name{extract_file_extension}
\alias{extract_file_extension}
\title{Extracts the extension of a file name}
\usage{
extract_file_extension(file_input)
}
\arguments{
\item{file_input}{Character string of a file name or path. Must only include
one ".", which is used as the separator.}
}
\description{
Extracts the extension of a file name
}
\details{
Extracts the file extension from a character string using "\\." as the
  separator.
}
