% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{extract_hobo_sn}
\alias{extract_hobo_sn}
\title{Extract hobo serial number from the data file}
\usage{
extract_hobo_sn(hobo_colnames)
}
\arguments{
\item{hobo_colnames}{Column names of the hobo file, as imported by
\code{ss_read_hobo_data()}.}
}
\value{
Returns the hobo serial number.
}
\description{
Extract hobo serial number from the data file
}
