% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{extract_hobo_units}
\alias{extract_hobo_units}
\title{Extract units from column names of hobo data}
\usage{
extract_hobo_units(hobo_dat)
}
\arguments{
\item{hobo_dat}{Data as read in by \code{ss_read_hobo_data()}.}
}
\value{
Returns a tibble of \code{variable} and \code{units} found in
  \code{hobo_dat}. Units are mg_per_l for dissolved oxygen and degree_c for
  temperature.
}
\description{
Extract units from column names of hobo data
}
