% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{extract_vemco_tz}
\alias{extract_vemco_tz}
\title{Extract the timezone of vemco timestamps}
\usage{
extract_vemco_tz(dat_colnames)
}
\arguments{
\item{dat_colnames}{Column names of the Vemco file.}
}
\value{
Returns a character string of the timezone indicated in the Timestamp
  column.
}
\description{
Extract the timezone of vemco timestamps
}
