% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_plot_variables_at_depth.R
\name{get_colour_palette}
\alias{get_colour_palette}
\title{Returns colour palette based on the unique values of
sensor_depth_at_low_tide_m}
\usage{
get_colour_palette(dat)
}
\arguments{
\item{dat}{Data to be plotted, **as returned by the function
\code{convert_to_tidydata()}. Must include the column \code{DEPTH}.**}
}
\value{
Returns a vector of hex colours from the viridis palette (Option D,
  direction = -1).
}
\description{
Returns colour palette based on the unique values of
sensor_depth_at_low_tide_m
}
\details{
Returns a discrete colour scale palette from the viridis package
  (Option D). If there are 6 or less unique values of
  \code{sensor_depth_at_low_tide_m}, the palette will have 6 colours. If
  there are more than 6 unique values of \code{sensor_depth_at_low_tide_m},
  the palette will have the number of colours equal to the number depths.
}
\author{
Danielle Dempsey
}
\concept{plot}
