% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{make_column_names}
\alias{make_column_names}
\title{Glue variable name and units to create column names}
\usage{
make_column_names(unit_table)
}
\arguments{
\item{unit_table}{Data frame including columns \code{variable} and
\code{units}, as returned from \code{extract_hobo_units()}.}
}
\value{
Data frame with column names in the form \code{variable_units}.
}
\description{
Glue variable name and units to create column names
}
