% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_hobo.R
\name{ss_compile_hobo_data}
\alias{ss_compile_hobo_data}
\title{Compile and format data from hobo or tidbit sensors}
\usage{
ss_compile_hobo_data(
  path,
  sn_table,
  deployment_dates,
  trim = TRUE,
  sensor_make = "hobo"
)
}
\arguments{
\item{path}{File path to the hobo or tidbit folder.}

\item{sn_table}{A data frame with three columns: \code{sensor_type},
\code{sensor_serial_number}, \code{depth}.}

\item{deployment_dates}{A data frame with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}

\item{sensor_make}{Character string indicating whether data to be compiled is
from hobo sensors (default) or tidbit sensors. Will be used to fill in the
\code{sensor_type} column of the output file. The raw file format is the
same for each type of sensor.}
}
\value{
Returns a tibble with the data compiled from each of the hobo or
  tidbit sensors.
}
\description{
Compile and format data from hobo or tidbit sensors
}
\details{
This code does not correct dissolved oxygen data for salinity.

  Exported hobo and tidbit data must be saved in a folder named hobo in csv
  format.

  All of the csv files in the hobo or tidbit folder will be compiled.

  The timestamp columns must be in the order "ymd IMS p", "Ymd IMS p", "Ymd
  HM", "Ymd HMS", "dmY HM", or "dmY HMS".
}
\examples{
path <- system.file("extdata", package = "sensorstrings")

sn_table <- data.frame(
  sensor_type = "hobo",
  sensor_serial_number = 10755220,
  depth = 2
)

deployment_dates <- data.frame(START = "2019-05-30", END = "2019-10-19")

hobo <- ss_compile_hobo_data(
  path,
  sn_table = sn_table,
  deployment_dates = deployment_dates
)
head(hobo)
}
