% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_vr2ar.R
\name{ss_compile_vr2ar_data}
\alias{ss_compile_vr2ar_data}
\title{Format temperature data from VR2AR deployment}
\usage{
ss_compile_vr2ar_data(path, sn_table, deployment_dates, trim = TRUE)
}
\arguments{
\item{path}{File path to the vr2ar folder. This folder should have one csv
file that was extracted using Fathom software. Other file types in the
folder will be ignored.}

\item{sn_table}{A data frame with three columns: \code{sensor_type},
\code{sensor_serial_number}, \code{depth}.}

\item{deployment_dates}{A data frame with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}
}
\value{
Returns a tibble with the data compiled from the file in path/vr2ar.
}
\description{
Compiles and formats temperature and depth data from VR2AR
  sensors exported from Fathom software.
}
\details{
The raw VR2AR data must be saved in a folder named vr2ar in csv
  format. Folder name is not case-sensitive.
}
\seealso{
Other compile: 
\code{\link{ss_compile_aquameasure_data}()},
\code{\link{ss_compile_deployment_data}()},
\code{\link{ss_compile_hobo_data}()},
\code{\link{ss_compile_hobo_ph_data}()},
\code{\link{ss_compile_vemco_data}()},
\code{\link{ss_read_log_old}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
