% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_log.R
\name{ss_convert_old_log}
\alias{ss_convert_old_log}
\title{Convert column names from old log into the new format}
\usage{
ss_convert_old_log(log)
}
\arguments{
\item{log}{Log file as read in from \code{ss_read_log()}.}
}
\value{
Returns log with columns names matching those as if it head been
  read in from
}
\description{
Convert column names from old log into the new format
}
