% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_coords_from_ddm_to_dd.R
\name{ss_coords_from_ddm_to_dd}
\alias{ss_coords_from_ddm_to_dd}
\title{Convert coordinates from degree-minute-decimal to decimal degree}
\usage{
ss_coords_from_ddm_to_dd(coords_ddm, west = TRUE)
}
\arguments{
\item{coords_ddm}{Vector of coordinate values in degree decimal minutes (xx
xx.xxxx, e.g., 45 21.651 or 64 2.063). These will be converted to decimal
degrees.}

\item{west}{Logical argument indicating if the coordinates are in the Western
hemisphere. If \code{TRUE} (the default), the returned longitudes will have
a negative sign.}
}
\value{
Returns a vector of coordinates in decimal-degree format.
}
\description{
To apply to multiple columns using \code{dplyr::across}:
}
\details{
1. Define variables to convert:

coords_ddm <- c("deployment_latitude_n_ddm", "deployment_longitude_w_ddm",
"retrieval_latitude_n_ddm", "retrieval_longitude_w_ddm")

2. Use the .names argument to specify new column names so that the original
columns are not over-written:

\code{dat <- dat_raw %>% mutate(across(any_of(coords_ddm),
~ss_coords_from_ddm_to_dd(.x), .names = "{str_remove(.col, '_w_ddm|_n_ddm')}"
))}
}
