% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_coords_to_dd.R
\name{ss_coords_to_decimal_degrees}
\alias{ss_coords_to_decimal_degrees}
\title{Convert coordiantes from degree-minute-decimal to decimal degree}
\usage{
ss_coords_to_decimal_degrees(dat)
}
\arguments{
\item{dat}{Data frame with columns \code{deployment_latitude_n_ddm} and
\code{deployment_longitude_w_ddm}. These will be converted to decimal
degrees if the corresponding values in columns \code{deployment_latitude}
and \code{deployment_longitude} are \code{NA}.}
}
\value{
Returns \code{dat} with NA values in columns
  \code{deployment_latitude} and \code{deployment_longitude} filled in based
  on the values in \code{deployment_latitude_n_ddm} and
  \code{deployment_longitude_w_ddm}.
}
\description{
Convert coordiantes from degree-minute-decimal to decimal degree
}
