% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_create_log_from_metadata.R
\name{ss_create_log_from_metadata}
\alias{ss_create_log_from_metadata}
\title{Writes a deployment log from the metadata tracking sheet}
\usage{
ss_create_log_from_metadata(
  output_path,
  station,
  deployment_date,
  path = NULL,
  sheet = "tracker",
  google_sheet = FALSE,
  google_sheet_link = NULL,
  to_title = TRUE
)
}
\arguments{
\item{output_path}{Path to the station deployment folder.}

\item{station}{Station name.}

\item{deployment_date}{Date of deployment as a character string in the format
"yyyy-mm-dd"}

\item{path}{Path to the metadata tracking sheet (including the file
name and extension). This must be a .xlsx file.}

\item{sheet}{Name of the tab with the deployment information.}

\item{google_sheet}{Logical argument indicating whether metadata is stored in
a google sheet. If \code{TRUE}, \code{path} is not required.}

\item{google_sheet_link}{Link the to metadata sheet on google drive. If
\code{NULL}, the New Brunswick metadata sheet will be imported.}

\item{to_title}{Logical argument indicating whether to convert the station
name to title case. Default is \code{TRUE}. Set to \code{FALSE} for
stations with capital letters within a word, e.g., McNabs Island.}
}
\value{
Returns deployment log in .csv format.
}
\description{
Writes a deployment log from the metadata tracking sheet
}
\details{
Imports the water quality metadata tracking sheet, filters for the
  station and date of interest, re-formats into the deployment log format,
  and exports to the Log folder.
}
