% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-plot.R
\name{ss_create_variable_labels_no_newline}
\alias{ss_create_variable_labels_no_newline}
\title{Create plot labels from variable names without new lines}
\usage{
ss_create_variable_labels_no_newline(dat)
}
\arguments{
\item{dat}{Data frame of Water Quality data with variables in long
format. Entries in \code{variable} column must be
\code{dissolved_oxygen_percent_saturation},
\code{dissolved_oxygen_uncorrected_mg_per_l}, \code{salinity_psu},
\code{sensor_depth_measured_m}, or \code{temperature_degree_c}.}
}
\value{
Returns \code{dat_long} with an addition column
  \code{variable_label}. \code{variable_label}
}
\description{
Create plot labels from variable names without new lines
}
