% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_dissolved_oxygen_correction.R
\name{ss_dissolved_oxygen_pressure_correction}
\alias{ss_dissolved_oxygen_pressure_correction}
\title{Pressure correction factor for dissolved oxygen measurements}
\usage{
ss_dissolved_oxygen_pressure_correction(dat_wide, sal = NULL, p_atm = NULL)
}
\arguments{
\item{dat_wide}{Data frame with at least one column:
\code{temperature_degree_C}. Corresponding pressure (atm) data may be
included in column \code{pressure_atm}. Additional columns will be ignored
and returned.}

\item{sal}{A single value of salinity (psu). This value must be specified if
there is no \code{salinity_psu} column in \code{dat_wide}. Default is
\code{sal = NULL}. Note: if \code{sal} is specified when there is a
\code{salinity_psu} column in \code{dat_wide}, the function will stop with
an error.}

\item{p_atm}{A single value of barometric pressure (atm). This value should
be specified if there is no \code{Pressure} column in \code{dat_wide}.
Default is \code{p_atm = NULL}. Note: if \code{p_atm} is specified when
there is a \code{pressure_atm} column in \code{dat_wide}, function will
stop with an error.}
}
\value{
Returns \code{dat_wide} with additional column(s), \code{F_p} and
  \code{pressure_atm}.
}
\description{
Pressure correction factor for dissolved oxygen measurements
}
\details{
Dissolved oxygen concentration (and partial pressure?) measurements
  should be corrected for atmospheric pressure if the pressure deviates from
  substantially from 1 atm (Benson & Krause, 1984).

  This function calculates pressure correction factor following Benson and
  Krause (1984), as suggested by USGS (2011). This is similar to what is
  described in the HOBO manual HOBO U26 Percent Saturation Calculation.pdf
  (except this function accounts for salinity in the water vapour
  calculation).

  Equation 24 in Benson & Krause 1984:

  \deqn{C_{p} = C_{star} * Pressure * (((1 - P_{wv} / Pressure) * (1 - theta
  * Pressure)) / ((1 - P_{wv}) * (1 - theta)))}

  \deqn{F_{p} = Pressure * (((1 - P_{wv} / Pressure) * (1 - theta *
  Pressure)) / ((1 - P_{wv}) * (1 - theta)))}

  \eqn{P_{wv}} is water vapour pressure, which depends on temperature and
  salinity and \eqn{theta} depends on the second virial coefficient of
  oxygen, and is calculated using temperature (see Table 2 of Benson and
  Krause 1984).

  \bold{References}

  Benson, Bruce B., Krause, Daniel, (1984), \emph{The concentration and
  isotopic fractionation of oxygen dissolved in freshwater and seawater in
  equilibrium with the atmosphere, Limnology and Oceanography}, 3, doi:
  10.4319/lo.1984.29.3.0620.

  USGS. \emph{Change to Solubility Equations for Oxygen in Water.} Technical
  Memorandum 2011.03. USGS Office of Water Quality, 2011.
}
\seealso{
Other Dissolved Oxygen: 
\code{\link{ss_dissolved_oxygen_salinity_correction}()}
}
\concept{Dissolved Oxygen}
