% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_export_county_files.R
\name{ss_export_county_files}
\alias{ss_export_county_files}
\title{Export data into a csv and/or rds file}
\usage{
ss_export_county_files(
  dat,
  county = "",
  output_path = NULL,
  export_csv = TRUE,
  export_rds = TRUE
)
}
\arguments{
\item{dat}{Dataframe to be exported, e.g., output from
\code{qc_assemble_county_data()}.}

\item{county}{County name. Will be appended with today's date as the name of
the output file.}

\item{output_path}{Path to where the assembled file(s) will be exported.
The default will export the csv file to the open_data folder and the rds
file to the processed_data/assembled_data folder (both on the R drive). For
custom \code{output_path} entries, the csv and rds files will be exported
to the same folder.}

\item{export_csv}{Logical argument indicating whether the data should be
  exported as a csv file. File name will be todays-date_county.csv. Default
  is \code{TRUE}.

  Note: \code{timestamp_utc} is converted to a character before exporting to
  remove UTC formatting (2018-12-23T05:00:00Z). When re-imported into R, the
  UTC timezone can be added using  \code{lubridate::force_tz()}.

  NAs are exported as blank cells at the request of the Open Data Portal.}

\item{export_rds}{Logical argument indicating whether the assembled data
should be exported as a *.rds file. File name will be
todays-date_county.rds. Default is \code{TRUE}.}
}
\value{
Exports \code{dat} as a csv and/or rds file.
}
\description{
Export data into a csv and/or rds file
}
\author{
Danielle Dempsey
}
