% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{ss_extract_hobo_units}
\alias{ss_extract_hobo_units}
\title{Extract units from column names of hobo data}
\usage{
ss_extract_hobo_units(hobo_dat)
}
\arguments{
\item{hobo_dat}{data as read in by \code{ss_read_hobo_data()}.}
}
\value{
placeholder
}
\description{
Extract units from column names of hobo data
}
