% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-plot.R
\name{ss_filter_dat_to_plot}
\alias{ss_filter_dat_to_plot}
\title{Filter data before plotting to zoom in on interesting features}
\usage{
ss_filter_dat_to_plot(
  dat,
  filter_to = c("start", "end", "custom"),
  period = "2 days",
  custom_start = NULL,
  custom_end = NULL
)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format, as exported from
\code{ss_compile_deployment_data()}. Must include columns timestamp_,
sensor_depth_at_low_tide_m, sensor_type, sensor-serial number, and the
variables to plot (e.g., temperature_degree_c).}

\item{filter_to}{Shortcut for specifying where to filter \code{dat} before
plotting. Options are "start", "end", or "custom". Default is "start".}

\item{period}{Character string that can be converted to a \code{lubridate}
period. Default is \code{"2 days"}.}

\item{custom_start}{Only required if \code{filter_to = "custom"}. POSIXct
object indicating where the filtered data will begin.}

\item{custom_end}{Only required if \code{filter_to = "custom"}. POSIXct
object indicating where the filtered data will end.}
}
\value{
Returns \code{dat} filtered to the specified dates.
}
\description{
Called by \code{ss_open_trimdates_app()}.
}
