% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_format_report_table.R
\name{ss_format_report_table}
\alias{ss_format_report_table}
\title{Format tables for summary report}
\usage{
ss_format_report_table(report_table)
}
\arguments{
\item{report_table}{Table to include in the summary report.}
}
\value{
Returns a flextable object that will render nicely in the report.
}
\description{
Format tables for summary report
}
