% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-misc.R
\name{ss_generate_depl_filepath}
\alias{ss_generate_depl_filepath}
\title{Generate file path and name for sensor string deployment data.}
\usage{
ss_generate_depl_filepath(dat, path = NULL, sub_folder = NULL, ext = "rds")
}
\arguments{
\item{dat}{Data frame of sensor string data in wide or long format. Must
include columns \code{county}, \code{station}, and \code{deployment_range}.}

\item{path}{Character string of the file path to the county folder.}

\item{sub_folder}{Character string of the sub-folder name (inside county
folder) where \code{dat} should be exported. Default is \code{sub-folder =
"new"}.}

\item{ext}{File extension. Default is \code{ext = "csv"}.}
}
\value{
A file path for exporting deployment data, including file name and
  extension.
}
\description{
Generate file path and name for sensor string deployment data.
}
