% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_leaflet_map.R
\name{ss_leaflet_station_map}
\alias{ss_leaflet_station_map}
\title{Interactive map of station location}
\usage{
ss_leaflet_station_map(dat)
}
\arguments{
\item{dat}{Data frame with columns \code{longitude}, \code{latitude},
\code{station}.}
}
\value{
A leaflet object.
}
\description{
Interactive map of station location
}
