% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{ss_make_column_names}
\alias{ss_make_column_names}
\title{temporary title}
\usage{
ss_make_column_names(unit_table)
}
\arguments{
\item{unit_table}{column variable and units}
}
\value{
dataframe with col names
}
\description{
temporary title
}
