% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_map_stations.R
\name{ss_map_stations}
\alias{ss_map_stations}
\title{Interactive map of station location(s)}
\usage{
ss_map_stations(dat)
}
\arguments{
\item{dat}{Data frame with columns \code{longitude}, \code{latitude},
\code{station}.}
}
\value{
A leaflet object.
}
\description{
Interactive map of station location(s)
}
\examples{
if (requireNamespace("leaflet", quietly = TRUE)) {
  stations <- data.frame(
    station = c("Borgles Island", "Birchy Head"),
    latitude = c(44.77241, 44.5701),
    longitude = c(-62.72608, -64.034383)
  )

  # the map tiles load when the map is viewed (needs internet)
  ss_map_stations(stations)
}
}
