% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_plot_variables_at_depth.R
\name{ss_plot_variables_at_depth}
\alias{ss_plot_variables_at_depth}
\title{Plot variables at depth}
\usage{
ss_plot_variables_at_depth(
  dat,
  measured_depth = TRUE,
  color_palette = NULL,
  legend_name = "Depth (m)",
  legend_position = "right"
)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format, as exported from
]code{ss_compile_deployment_data()}.}

\item{measured_depth}{Logical variable indicating whether to include
\code{sensor_depth_measured_m} as a plotted variable.}

\item{color_palette}{Optional vector of hex colors onto which depth will be
mapped.}

\item{legend_name}{Name for the depth legend. Default is \code{legend_name =
"Depth (m)"}.}

\item{legend_position}{Position for the depth legend. Default is
\code{legend.position = "right"}.}
}
\value{
Returns a ggplot object of ocean variables plotted over time and
  coloured by sensor depth.
}
\description{
Plot variables at depth
}
