% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_log_simple.R
\name{ss_read_log_simple}
\alias{ss_read_log_simple}
\title{Read in deployment log}
\usage{
ss_read_log_simple(path, parse = TRUE)
}
\arguments{
\item{path}{File path to the Log folder, or full path to the log file,
include file name and extension.}

\item{parse}{Logical argument indicating whether to parse log into a list
used by ss_compile functions.}
}
\value{
Returns a a data frame of the deployment metadata. Option to parse
  the information into a list using \code{ss_parse_log()}.
}
\description{
Read in deployment log
}
\details{
The log must be saved in a folder called Log in .csv, .xlsx or .xls
  format, and must include the following columns:

  \code{Deployment_Waterbody}: waterbody where string was deployed

  \code{Location_Description}: the station name

  \code{Lease#}: If located on an aquaculture site, the lease number (NA
  otherwise)

  \code{Deployment}: deployment date, in the order "Ymd"

  \code{Retrieval}: retrieval date, in the order "Ymd"

  \code{Logger_Latitude}: The latitude at which the string was deployed

  \code{Logger_Longitude} The longitude at which the string was deployed
  (must be a negative value)

  \code{Logger_Model} The type of sensor; see below for options

  \code{Serial#} The sensor serial number

  \code{Sensor_Depth}: Depth at which the sensor was deployed

  All other columns will be ignored.

  Entries in the \code{Logger_Model} column must include the string
  "aquameasure", "hobo", "tidbit", or "vr2ar" (not case sensitive).

  The function will stop with an Error if there is more than one eligible
  file (csv, .xlsx or .xls) in the Log folder.

  The function will stop with an Error if there if the
  \code{Logger_Longitude} is a positive value.

  A Warning message is printed to the console when the function does not
  recognize a sensor in the log.

  A message is printed to the console when hobo, aquameasure, or vemco
  sensors are not found in the log.

  A message is printed to the console if there is more than one unique entry
  in \code{Deployment_Waterbody}, \code{Location_Description},
  \code{Deployment}, \code{Retrieval}, \code{Logger_Latitude}, or
  \code{Logger_Longitude}.
}
