% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_nsdfa_metadata.R
\name{ss_read_nsdfa_metadata}
\alias{ss_read_nsdfa_metadata}
\title{Import TempMetaData tab from the NSDFA tracking sheet}
\usage{
ss_read_nsdfa_metadata(path)
}
\arguments{
\item{path}{Path to the NSDFA tracking sheet (include file name and
extension).}
}
\value{
Returns data frame of NSDFA tracking sheet sensor string metadata.
}
\description{
Import TempMetaData tab from the NSDFA tracking sheet
}
\details{
Reads in the TempMetaData tab from the NSDFA tracking sheet and
  corrects known errors (e.g., standardizes station and waterbody spellings,
  fixes deployment dates, etc.).
}
\examples{
# A small synthetic version of the NSDFA tracking sheet (TempMetaData tab).
# The function fixes misspelled stations and waterbodies, removes times from
# the dates, and makes positive longitudes negative. tidyr gives an
# "Expected 2 pieces" warning for dates that have no time; this is expected.

nsdfa <- ss_read_nsdfa_metadata(
system.file("extdata/nsdfa_tracking_example.xlsx",  package = "sensorstrings"))

nsdfa[, c("County", "Waterbody", "Station_Name", "Depl_Date", "Recv_Date", "Depl_Lon")]
}
