% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_reformat_old_data.R
\name{ss_reformat_old_data}
\alias{ss_reformat_old_data}
\title{Convert old data structure to sensorstrings format}
\usage{
ss_reformat_old_data(dat)
}
\arguments{
\item{dat}{Water quality data in the old format.}
}
\value{
\code{dat} formatted for use with \code{sensorstrings} and
  \code{qaqcmar}.
}
\description{
Convert old data structure to sensorstrings format
}
