% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-plot.R
\name{ss_xaxis_breaks}
\alias{ss_xaxis_breaks}
\title{Returns nice major an
d minor breaks and label format based on timespan
 of the data}
\usage{
ss_xaxis_breaks(dat)
}
\arguments{
\item{dat}{Data frame with at least one column: \code{timestamp_utc} (POSIXct).}
}
\value{
Returns a dataframe with 1 observation of 3 variables
 \code{date_breaks_major}, \code{date_breaks_minor},
 \code{date_labels_format}.
}
\description{
Returns nice major an
d minor breaks and label format based on timespan
 of the data
}
\details{
If the timespan in the \code{timestamp_utc} column of \code{dat} is
 less than or equal to 60 days, the major and minor breaks will be every 2
 weeks; if time span is greater than 60 days and less than or equal to 120
 days, the major and minor breaks will be every month; if time span is
 greater than 120 days and less than or equal to 240 days, the major breaks
 will be 2 months and the minor breaks will be 1 month; if the time span is
 greater than 240 days, the major breaks will be 4 months and the minor
 breaks will be 1 month.
}
