% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{trim_data}
\alias{trim_data}
\title{Trim data to specified start and end dates.}
\usage{
trim_data(dat, start_date, end_date)
}
\arguments{
\item{dat}{Data.frame with column including the string "timestamp"}

\item{start_date}{POSIXct/POSIXt value of the first good timestamp.}

\item{end_date}{POSIXct/POSIXt value of the last good timestamp.}
}
\value{
Returns dat trimmed.
}
\description{
4 hours adde to end_date to account for AST (e.g., in case the sensor was
retrieved after 20:00 AST, which is 00:00 UTC **The next day**)
}
