% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_timestamp_to_datetime}
\alias{convert_timestamp_to_datetime}
\title{convert_timestamp_to_datetime()}
\usage{
convert_timestamp_to_datetime(dat)
}
\arguments{
\item{dat}{Data.frame with column \code{timestamp_} that has timestamps as
character values.}
}
\description{
convert_timestamp_to_datetime()
}
\details{
Convert the timestamp_ column to a POSIXct object.
}
