% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{extract_deployment_dates}
\alias{extract_deployment_dates}
\title{Extract deployment dates}
\usage{
extract_deployment_dates(deployment_dates)
}
\arguments{
\item{deployment_dates}{Data.frame with start and end dates of the deployment
in the form yyyy-mm-dd. Two columns: \code{START} and \code{END}.}
}
\description{
Extract deployment dates
}
