% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{extract_file_extension}
\alias{extract_file_extension}
\title{Extracts the extension of a file name}
\usage{
extract_file_extension(file_name)
}
\arguments{
\item{file_name}{Character string of a file name. Must only include one ".",
which is used as the seprator.}
}
\description{
Extracts the extension of a file name
}
\details{
Extracts the file exension from a character string using //. as the
 separator.
}
