% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{extract_hobo_sn}
\alias{extract_hobo_sn}
\title{Extract HOBO serial number from the data file}
\usage{
extract_hobo_sn(hobo_colnames)
}
\arguments{
\item{hobo_colnames}{Column names of the HOBO file, as imported by
\code{ss_read_hobo_data}.}
}
\value{
Returns a character string of the HOBO serial number.
}
\description{
Extract HOBO serial number from the data file
}
