% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-plot.R
\name{filter_dat_to_plot}
\alias{filter_dat_to_plot}
\title{Filter data before plotting to zoom in on interesting features}
\usage{
filter_dat_to_plot(
  dat,
  filter_to = c("start", "end", "custom"),
  period = "2 days",
  custom_start = NULL,
  custom_end = NULL
)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format, as exported from
\code{ss_compile_deployment_data()}.}

\item{filter_to}{Shortcut for specifying where to filter \code{dat} before
plotting. Options are "start", "end", or "custom".}

\item{period}{Character string that can be convert to a \code{lubridate}
period.}

\item{custom_start}{Only required if \code{filter_to = "custom"}. POSIXct
object indicating where the filtered data will begin.}

\item{custom_end}{Only required if \code{filter_to = "custom"}. POSIXct
object indicating where the filtered data will end.}
}
\value{
Returns a plotly object of the filtered data coloured by depth and
  faceted by variable.
}
\description{
Filter data before plotting to zoom in on interesting features
}
