% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{set_up_compile}
\alias{set_up_compile}
\title{Temp}
\usage{
set_up_compile(path, sn_table, deployment_dates, sensor_make, verbose = FALSE)
}
\arguments{
\item{path}{temp}

\item{sn_table}{temp}

\item{deployment_dates}{temp}

\item{sensor_make}{temp}

\item{verbose}{temp}
}
\value{
temp
}
\description{
Temp
}
