% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{set_up_compile}
\alias{set_up_compile}
\title{Set up parameters, Errors, and Warnings for the \code{compile_**} functions}
\usage{
set_up_compile(path, sn_table, deployment_dates, sensor_make)
}
\arguments{
\item{path}{File path to the folder with the aquameasure, hobo, or vemco
folder.}

\item{sn_table}{A data frame with three columns: \code{sensor}, code{serial},
\code{depth}, as returned by \code{ss_read_log()}.}

\item{deployment_dates}{A dataframe with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{sensor_make}{Make of the sensor to be compiled. Should match the name
of the folder where the raw data files are saved and be found in the
\code{sensor} column in \code{sn_table}. Most common entries will be
"aquameasure", "hobo", or "vemco".}
}
\value{
Returns a list of parameters used in the \code{compile_**} functions.
  Returns Errors and Warnings if the expected files are not found in
  \code{folder}.
}
\description{
Set up parameters, Errors, and Warnings for the \code{compile_**} functions
}
