% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compile.R
\name{set_up_compile}
\alias{set_up_compile}
\title{Set up parameters, Errors, and Warnings for the \code{compile_**} functions}
\usage{
set_up_compile(path, sn_table, deployment_dates, sensor_make)
}
\arguments{
\item{path}{File path to the folder with the aquameasure, hobo, or vemco
folder.}

\item{sn_table}{A data frame with three columns: \code{sensor_type},
\code{sensor_serial_number}, \code{depth}.}

\item{deployment_dates}{A data frame with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{sensor_make}{Make of the sensor to be compiled. Should match the name
of the folder where the raw data files. Most common entries will be
"aquameasure", "hobo", or "vemco".}
}
\value{
Returns a list of parameters used in the \code{compile_**} functions:
  final path to the folder of interest, deployment dates, vector of files in
  the folder, \code{sn_table} filtered for \code{sensor_make}.
}
\description{
Set up parameters, Errors, and Warnings for the \code{compile_**} functions
}
\details{
An extra column named \code{sensor_type} is added to
  \code{sn_table}, with values of \code{sensor_make}. The \code{sensor_type}
  column is what is used in the compiled data (e.g., "aquameasure", "hobo",
  "vr2ar" vs "HOBO PRO V2" or "aquaMeasure DOT").

  Returns Errors and Warnings if the expected files are not found on the
  \code{path}.
}
