% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_assemble_region_datasets.R
\name{ss_assemble_region_data}
\alias{ss_assemble_region_data}
\title{Assemble region data}
\usage{
ss_assemble_region_data(prov = "ns", folder)
}
\arguments{
\item{prov}{Character string indicating which province the data to be
assembled is from. Options are "ns" (the default) and "nb". This dicates
the file path for where the data will be exported on the CMAR R drive.}

\item{folder}{Name of the folder where the rds files are saved.}
}
\value{
Returns a data.frame with data from all deployments in folder.
}
\description{
Assemble region data
}
