% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_check_station_location.R
\name{ss_check_station_drift}
\alias{ss_check_station_drift}
\title{Check station deployment and retrieval locations are within acceptable
distance}
\usage{
ss_check_station_drift(
  log_coords,
  log_crs = 4617,
  max_drift = 100,
  return_drift = FALSE
)
}
\arguments{
\item{log_coords}{Data frame with five columns: \code{station},
\code{latitude} and \code{longitude} (the coordinates recorded at
deployment) and retrieval_latitude and retrieval_longitude (the coordinates
recorded during retrieval).}

\item{log_crs}{The crs of the coordinates in \code{log_coords}.}

\item{max_drift}{The maximum acceptable distance between the deployment and
retrieval coordinates in metres.}

\item{return_drift}{Logical argument indicating whether to return the drift
distance.}
}
\value{
Logical value. Returns \code{TRUE} if the distance between deployment
  and retrieval coordinates is less than \code{max_drift}. Returns
  \code{FALSE} and a Warning if the distance is greater.
}
\description{
Check station deployment and retrieval locations are within acceptable
distance
}
