% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_check_station_location.R
\name{ss_check_station_in_ocean}
\alias{ss_check_station_in_ocean}
\title{Check station location is in the ocean (not on land)}
\usage{
ss_check_station_in_ocean(log_coords, log_crs = 4617, coast_shp = NULL)
}
\arguments{
\item{log_coords}{Data frame with three columns: \code{station},
\code{latitude} and \code{longitude}, i.e., with the station coordinates as
recorded in the deployment log.}

\item{log_crs}{The crs of the coordinates in \code{log_coords}.}

\item{coast_shp}{Shapefile of the coast for comparison to the station
location. The resolution of the shapefile can impact the results of this
test. Coarse resolution may result in false positives for overlapping with
land. If \code{NULL}, must be connected to the Perennia R drive.}
}
\value{
Logical value. Returns \code{TRUE} if the log coordinates are in the
  ocean. Returns \code{FALSE} and a Warning if the coordinates overlap with
  land.
}
\description{
Check station location is in the ocean (not on land)
}
