% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_check_station_location.R
\name{ss_check_station_radius}
\alias{ss_check_station_radius}
\title{Check station location is within buffer of official coordinates}
\usage{
ss_check_station_radius(
  log_coords,
  log_crs = 4617,
  station_coords = NULL,
  station_radius = 500
)
}
\arguments{
\item{log_coords}{Data frame with three columns: \code{station},
\code{latitude} and \code{longitude}, i.e., with the station coordinates as
recorded in the deployment log.}

\item{log_crs}{The crs of the coordinates in \code{log_coords}.}

\item{station_coords}{Data frame with two columns: \code{latitude} and
\code{longitude}, the "official" station coordinates. If \code{NULL}, the
station coordinates will be looked up in the AREA INFO tracking sheet.}

\item{station_radius}{Acceptable radius around the official station
coordinates (in metres). Stations outside of this radius should be renamed.
Default is 500 m.}
}
\value{
Logical value. Returns \code{TRUE} if the log coordinates are within
  the buffer radius. Returns \code{FALSE} and a Warning if the coordinates
  are outside the buffer.
}
\description{
Check station location is within buffer of official coordinates
}
