% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_aquameasure.R
\name{ss_compile_aquameasure_data}
\alias{ss_compile_aquameasure_data}
\title{Compiles data from aquameasure sensors}
\usage{
ss_compile_aquameasure_data(path, sn_table, deployment_dates, trim = TRUE)
}
\arguments{
\item{path}{File path to the hobo folder.}

\item{sn_table}{A data frame with three columns: \code{sensor_type},
code{sensor_serial_number}, \code{depth}.}

\item{deployment_dates}{A data frame with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}
}
\value{
Returns a tibble with the data compiled from each of the aquameasure
  files in path/aquameasure.
}
\description{
Compiles and formats temperature, dissolved oxygen, salinity,
  and/or device depth data from aquameasure sensors.
}
\details{
The raw aquameasure data must be saved in a folder named aquameasure
  in csv format. Folder name is not case-sensitive.

  Rows with \code{undefined} and \code{... (time not set)} values in the
  \code{Timestamp(UTC)} column are filtered out.

  The timestamp columns must be in the order "ymd IMS p", "Ymd IMS p", "Ymd
  HM", "Ymd HMS", "dmY HM", or "dmY HMS".

  MOVE THIS TO QAQCMAR???

  Negative Dissolved Oxygen values are converted to \code{NA}.

  "ERR" values are converted to \code{NA}.
}
\seealso{
Other compile: 
\code{\link{ss_compile_deployment_data}()},
\code{\link{ss_compile_hobo_data}()},
\code{\link{ss_compile_vemco_data}()},
\code{\link{ss_read_log}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
