% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_compile_deployment_data.R
\name{ss_compile_deployment_data}
\alias{ss_compile_deployment_data}
\title{Compile Water Quality data from a sensor string deployment}
\usage{
ss_compile_deployment_data(
  path,
  trim = TRUE,
  ignore_sensors = NULL,
  depth_override = NULL
)
}
\arguments{
\item{path}{File path to the log, aquameasure, hobo, hobo_ph, vdat,
and/or vemco folders.}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}

\item{ignore_sensors}{Vector of sensor serial numbers for sensors that are in
the deployment log, but should NOT be compiled (e.g., data file missing).}

\item{depth_override}{An optional character string indicating which depth
variable to compile. In some files (e.g., Borgles Island 2018-02-28), there
is only one "Seawater depth" observation, but a full deployment of "Average
seawater depth" observations. In this case, force the code to compile the
average seawater depth with \code{depth_override = "Average seawater
depth"}.}
}
\value{
Returns a tibble of data from a sensor string deployment.
}
\description{
Compile Water Quality data from a sensor string deployment
}
\details{
Reads the deployment log and then calls the
  \code{ss_compile_*_data()}, functions and returns the results in a single
  data frame.

  Data files must be in a folder named after the sensor manufacturer and/or
  model (e.g., aquameasure, hobo, hobo_ph, vemco, vdat). Folder names
  are not case sensitive. These folders must be in the same folder.

  VR2AR data processed with the Fathom software must be in a folder called
  vdat.

  Columns with deployment details are added (e.g., county, waterbody,
  latitude, longitude, station, lease, string_configuration).
}
\examples{
#path to the log, aquameasure, hobo, and vemco folders
path <- system.file("extdata", package = "sensorstrings")

dat <- ss_compile_deployment_data(path)
head(dat)

# leave the data untrimmed, and skip one of the sensors in the log
dat_untrimmed <- ss_compile_deployment_data(
  path, trim = FALSE, ignore_sensors = 547109
)
unique(dat_untrimmed$sensor_type)
}
\seealso{
Other compile:
\code{\link[=ss_compile_aquameasure_data]{ss_compile_aquameasure_data()}},
\code{\link[=ss_compile_vdat_data]{ss_compile_vdat_data()}},
\code{\link[=ss_compile_vemco_data]{ss_compile_vemco_data()}}
}
\author{
Danielle Dempsey
}
\concept{compile}
