% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_compile_deployment_data.R
\name{ss_compile_deployment_data}
\alias{ss_compile_deployment_data}
\title{Compile aquameasure, hobo, and vemco data from a single deployment}
\usage{
ss_compile_deployment_data(
  path,
  path_config = NULL,
  trim = TRUE,
  ignore_sensors = NULL
)
}
\arguments{
\item{path}{File path to the log, aquameasure, hobo, tidbit, and/or vemco
folders.}

\item{path_config}{File path (including name and extension) of the water
  quality configuration table. Default is
  water_quality_configuration_table.xlsx in the tracking_sheets folder. Data
  must be stored in an xlsx file in a tab named "deployments". Columns must
  be Waterbody, Station_Name, and Depl_Date. One and only one row must match
  the waterbody, station, and deployment date recorded in the Log. Options
  for the configuration column are sub-surface buoy, surface buoy, attached
  to gear, attached to fixed structure, floating dock, and unknown. Use of
  the "unknown" entry is discouraged. HINT: for optimal speed, make sure the
  table is NOT filtered in the excel file.

  The \code{path_config} argument is ignored for new deployment logs that
  have an appropriate entry in the Configuration column.}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}

\item{ignore_sensors}{Vector of sensor serial numbers for sensors that are in
the deployment log, but should NOT be compiled (e.g., data file missing).}
}
\value{
Returns a data frame of data from a single sensor string deployment.
}
\description{
Compile aquameasure, hobo, and vemco data from a single deployment
}
\details{
Reads the deployment log and then calls
  \code{ss_compile_aquameasure_data()}, \code{ss_compile_hobo_data()}, and
  \code{ss_compile_vemco_data()} and returns the results in a single data
  frame.

  aquameasure data must be in a folder named aquameasure, hobo data must be
  in a folder named hobo, tidbit data must be in a folder named tidbit, and
  vemco data must be in a folder name vemco (folder names are not case
  sensitive). The aquameasure, hobo, tidbit, and vemco folders must be in the
  same folder.

  Adds location and mooring columns
}
\seealso{
Other compile: 
\code{\link{ss_compile_aquameasure_data}()},
\code{\link{ss_compile_hobo_data}()},
\code{\link{ss_compile_vemco_data}()},
\code{\link{ss_read_log}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
