% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_hobo.R
\name{ss_compile_hobo_data}
\alias{ss_compile_hobo_data}
\title{Compile and format data from Hobo and TidbiT sensors}
\usage{
ss_compile_hobo_data(path, sn_table, deployment_dates, trim = TRUE)
}
\arguments{
\item{path}{File path to the Hobo folder.}

\item{sn_table}{A data frame with three columns: \code{sensor}, code{serial},
\code{depth}, as returned by \code{ss_read_log()}.}

\item{deployment_dates}{A dataframe with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}
}
\value{
Returns a tibble with the data compiled from each of the HOBO and
  TidbiT sensors in path/Hobo.
}
\description{
Can handle temperature and dissolved oxygen data.

  Need to decide if the DO data will be corrected for salinity!
}
\details{
The exported hobo data must be saved in a folder named Hobo in csv
  format. Folder name is not case-sensitive.

  All of the csv files in the Hobo folder will be compiled. The name of each
  file must be the serial number of the sensor,

  The timestamp columns must be in the order "ymd IMS p", "Ymd IMS p", "Ymd
  HM", "Ymd HMS", "dmY HM", or "dmY HMS".
}
\seealso{
Other compile: 
\code{\link{ss_compile_aquameasure_data}()},
\code{\link{ss_compile_deployment_data}()},
\code{\link{ss_compile_vemco_data}()},
\code{\link{ss_read_log}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
