% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_hobo_ph.R
\name{ss_compile_hobo_ph_data}
\alias{ss_compile_hobo_ph_data}
\title{Compile and format data from hobo pH sensors (MX2501)}
\usage{
ss_compile_hobo_ph_data(
  path,
  sn_table,
  deployment_dates,
  tz_check = FALSE,
  trim = TRUE
)
}
\arguments{
\item{path}{File path to the hobo_ph folder.}

\item{sn_table}{A data frame with three columns: \code{sensor_type},
\code{sensor_serial_number}, \code{depth}.}

\item{deployment_dates}{A data frame with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{tz_check}{Logical argument indicating whether to check if the timestamp
column is in utc. If \code{TRUE}, a warning will be printed if the timezone
is NOT utc.}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}
}
\value{
Returns a tibble with the data compiled from each of the hobo pH
  sensors.
}
\description{
Compile and format data from hobo pH sensors (MX2501)
}
\details{
Exported data must be saved in a folder named hobo_ph in csv format.

  All of the csv files in the hobo_ph folder will be compiled.

  The timestamp column must be in the order "mdY HMS", “Ymd HMS”, or “Ymd”.

  If the timestamp is recorded in AST/ADT, it will be converted to UTC.
}
\examples{
# path holds a folder called hobo_ph
path <- system.file("testdata", package = "sensorstrings")

sn_table <- data.frame(
  sensor_type = "hobo ph",
  sensor_serial_number = 22058687,
  depth = 1
)

deployment_dates <- data.frame(START = "2025-03-08", END = "2025-03-09")

hobo_ph <- ss_compile_hobo_ph_data(
  path,
  sn_table = sn_table,
  deployment_dates = deployment_dates
)
head(hobo_ph)
}
