% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_vdat.R
\name{ss_compile_vdat_data}
\alias{ss_compile_vdat_data}
\title{Format temperature, depth, and tilt data from VR2AR (vdat) deployment}
\usage{
ss_compile_vdat_data(path, sn_table, deployment_dates, trim = TRUE)
}
\arguments{
\item{path}{File path to the vdat folder. This folder should have one csv
file that was extracted using Fathom software. Other file types in the
folder will be ignored.}

\item{sn_table}{A data frame with three columns: \code{sensor_type},
\code{sensor_serial_number}, \code{depth}.}

\item{deployment_dates}{A data frame with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}
}
\value{
Returns a tibble with the data compiled from the file in path/vdat.
}
\description{
Compiles and formats temperature, depth, and tilt data from
  VR2AR data offloaded in the new .vdat format and exported with Fathom
  software.
}
\details{
The raw VR2AR data must be saved in a folder named vdat in csv
  format. Folder name is not case-sensitive.

  The device timestamp is used here. This may drift over time. Corrected time
  stamp is included in the .vdat file, but the date is not always appended.
}
\examples{
# full path to a VR2AR file processed with Fathom (vdat)
path <- system.file(
  "testdata", "vdat", "VR2AR-69_555438.csv", package = "sensorstrings"
)

sn_table <- data.frame(
  sensor_type = "VR2AR-69",
  sensor_serial_number = 555438,
  depth = 5
)

deployment_dates <- data.frame(START = "2019-05-30", END = "2019-10-19")

vdat <- ss_compile_vdat_data(
  path,
  sn_table = sn_table,
  deployment_dates = deployment_dates,
  trim = FALSE
)
head(vdat)
}
\seealso{
Other compile:
\code{\link[=ss_compile_aquameasure_data]{ss_compile_aquameasure_data()}},
\code{\link[=ss_compile_vemco_data]{ss_compile_vemco_data()}}
}
\author{
Danielle Dempsey
}
\concept{compile}
