% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_vemco.R
\name{ss_compile_vemco_data}
\alias{ss_compile_vemco_data}
\title{Format temperature, depth, and tilt data from Vemco deployment}
\usage{
ss_compile_vemco_data(
  path,
  sn_table,
  deployment_dates,
  trim = TRUE,
  depth_override = NULL
)
}
\arguments{
\item{path}{File path to the vemco folder. This folder should have one csv
file that was extracted using Vue software. Other file types in the folder
will be ignored.}

\item{sn_table}{A data frame with three columns: \code{sensor_type},
\code{sensor_serial_number}, \code{depth}.}

\item{deployment_dates}{A data frame with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}

\item{depth_override}{An optional character string indicating which depth
variable to compile. In some files (e.g., Borgles Island 2018-02-28), there
is only one "Seawater depth" observation, but a full deployment of "Average
seawater depth" observations. In this case, force the code to compile the
average seawater depth with \code{depth_override = "Average seawater
depth"}.}
}
\value{
Returns a tibble with the data compiled from the file in path/vemco.
}
\description{
Compiles and formats temperature, seawater depth, and device
  tilt data from VR2AR sensors processed with Vue software.
}
\details{
The raw vemco data must be saved in a folder named vemco in csv
  format. Folder name is not case-sensitive.

  If there are "Temperature" entries in the Description column, these will be
  extracted and compiled. If there are no "Temperature" entries, but there
  are "Average temperature" entries, these will be extracted and compiled.
  Otherwise, the function will stop with an error message.

  If there are "Seawater depth" entries in the Description column, these will
  be extracted and compiled. If there are no "Seawater depth" entries, but
  there are "Average seawater depth" entries, these will be extracted and
  compiled. Otherwise, the function will stop with an error message.  (See
  argument \code{depth_override} for exception.)
}
\examples{
path <- system.file("extdata", package = "sensorstrings")

sn_table <- data.frame(
  sensor_type = "VR2AR",
  sensor_serial_number = 547109,
  depth = 15
)

deployment_dates <- data.frame(START = "2019-05-30", END = "2019-10-19")

vemco <- ss_compile_vemco_data(
  path,
  sn_table = sn_table,
  deployment_dates = deployment_dates
)
head(vemco)
}
\seealso{
Other compile:
\code{\link[=ss_compile_aquameasure_data]{ss_compile_aquameasure_data()}},
\code{\link[=ss_compile_deployment_data]{ss_compile_deployment_data()}},
\code{\link[=ss_compile_vdat_data]{ss_compile_vdat_data()}}
}
\author{
Danielle Dempsey
}
\concept{compile}
