% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_vemco.R
\name{ss_compile_vemco_data}
\alias{ss_compile_vemco_data}
\title{Format temperature data from Vemco deployment}
\usage{
ss_compile_vemco_data(path, sn_table, deployment_dates, trim = TRUE)
}
\arguments{
\item{path}{File path to the Vemco folder. This folder should have one csv
file that was extracted using Vue software. Other file types in the folder
will be ignored.}

\item{sn_table}{A data frame with three columns: \code{sensor}, code{serial},
\code{depth}, as returned by \code{ss_read_log()}.}

\item{deployment_dates}{A dataframe with two columns. The first column holds
the deployment date (a Date object, POSIXct object, or character string in
the order year, month, day), and the second column holds the retrieval date
(a Date object, POSIXct object, or character string in the order year,
month, day).}

\item{trim}{Logical value indicating whether to trim the data to the dates
specified in \code{deployment_dates}. (Note: four hours are added to the
retrieval date to account for AST, e.g., in case the sensor was retrieved
after 20:00 AST, which is 00:00 UTC the next day.) Default is \code{trim =
TRUE}.}
}
\value{
Returns a tibble with the data compiled from the file in path/vemco.
}
\description{
Compiles and formats temperature and seawater depth data from
  VR2AR sensors.
}
\details{
The raw Vemco data must be saved in a folder named Vemco in csv
  format. Folder name is not case-sensitive.

  If there are "Temperature" entries in the Description column, these will be
  extracted and compiled. If there are no "Temperature" entries, but there
  are "Average temperature" entries, these will be extracted and compiled.
  Otherwise, the function will stop with an error message.
}
\seealso{
Other compile: 
\code{\link{ss_compile_aquameasure_data}()},
\code{\link{ss_compile_deployment_data}()},
\code{\link{ss_compile_hobo_data}()},
\code{\link{ss_read_log}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
