% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_convert_depth_to_ordered_factor.R
\name{ss_convert_depth_to_ordered_factor}
\alias{ss_convert_depth_to_ordered_factor}
\title{Converts values in the depth column to an ordered factor}
\usage{
ss_convert_depth_to_ordered_factor(dat)
}
\arguments{
\item{dat}{Data frame that includes the column
\code{sensor_depth_at_low_tide_m}.}
}
\value{
Returns \code{dat}, with the \code{sensor_depth_at_low_tide_m} column
  converted to an ordered factor. The shallowest depth will have a factor
  value of 1; the deepest depth will have the largest factor value.
}
\description{
This function simplifies converting the
  sensor_depth_at_low_tide_m column from class \code{numeric}, \code{factor},
  or \code{character} to an ordered factor, which is preferred for using the
  function \code{ss_ggplot_variables()}.
}
\examples{
dat <- data.frame(
  sensor_depth_at_low_tide_m = c(10, 2, 5, 2),
  temperature_degree_c = c(8, 15, 12, 14)
)

dat_ordered <- ss_convert_depth_to_ordered_factor(dat)

dat_ordered
levels(dat_ordered$sensor_depth_at_low_tide_m)
}
\author{
Danielle Dempsey
}
\concept{format}
