% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_convert_depth_to_ordered_factor.R
\name{ss_convert_depth_to_ordered_factor}
\alias{ss_convert_depth_to_ordered_factor}
\title{Converts values in the DEPTH column to an ordered factor}
\usage{
ss_convert_depth_to_ordered_factor(dat)
}
\arguments{
\item{dat}{Data in tidy format, as returned by the function
\code{convert_to_tidydata()}. Must include at least one column:
\code{DEPTH} (factor, character, or numeric).}
}
\value{
Returns \code{dat.tidy}, with the \code{DEPTH} column converted to an
  ordered factor.
}
\description{
This function simplifies converting the DEPTH column from class
  \code{numeric}, \code{factor}, or \code{character} to an ordered factor,
  which is preferred for using the function \code{plot_variables_at_depth()}.
}
\details{
To use the function \code{plot_variables_at_depth()}, \code{DEPTH}
  must be a factor because \code{DEPTH} is mapped to a discrete color scale.
  For the legend to display correctly (e.g. depths arranged from shallow to
  deep), \code{DEPTH} must be an \emph{ordered} factor.

  The \code{DEPTH} column may be values that can be coerced to class
  \code{numeric}, qualitative descriptors, or a combination of both. A
  different level will be assigned to each of the unique values of
  \code{DEPTH}. This could cause some confusion when \code{DEPTH} includes
  numeric and qualitative values, for example because the level (colour) for
  "Surface" will not the same as the level for "2 m").

  Qualitative \code{DEPTH} values will be standardized to "Surface",
  "Sub-surface", "Middle", and "Bottom". Accepted qualitative \code{DEPTH}
  values are "Bottom", "bottom", "Middle", "middle", "Sub surface",
  "Sub-Surface", Sub-surface", "sub surface", "sub-surface", "Sub", "sub",
  "Surface", and "surface". Other values will cause the function to stop with
  an error message.

  For numeric depths: the smallest value is assigned to the first level, the
  second smallest value is assigned to the second level, and so on.

  For qualitative depths: the assigned order is "Surface" < "Sub-surface" <
  "Middle" < "Bottom".
}
\author{
Danielle Dempsey
}
\concept{format}
