% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_create_template.R
\name{ss_create_template}
\alias{ss_create_template}
\title{Create file to compile and trim sensor string data}
\usage{
ss_create_template(
  station = NULL,
  depl_date = NULL,
  initials = "",
  path = NULL
)
}
\arguments{
\item{station}{Character string of the station name. Will be converted to
lower case, and all spaces will be replaced with an underscore.}

\item{depl_date}{Character string of the deployment data in the order
yyyy-mm-dd.}

\item{initials}{Initials of Data Analyst. Default is "".}

\item{path}{Path to folder where file will be created. Default is the CMAR R
drive in folder << station >> / << station_depl_date >>. Must be connected
to the Perennia VPN.}
}
\value{
Invisibly returns the path to the template file. Creates the file if
  it doesn't already exist. Otherwise, returns a Warning and leaves the file
  unchanged.
}
\description{
Create file to compile and trim sensor string data
}
