% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-plot.R
\name{ss_create_variable_labels}
\alias{ss_create_variable_labels}
\title{Create plot labels from variable names}
\usage{
ss_create_variable_labels(dat)
}
\arguments{
\item{dat}{Data frame of Water Quality data with variables in long
format.}
}
\value{
Returns \code{dat_long} with an addition column
  \code{variable_label}. \code{variable_label}
}
\description{
Create plot labels from variable names
}
