% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-plot.R
\name{ss_get_colour_palette}
\alias{ss_get_colour_palette}
\title{Returns colour palette based on the unique values of
sensor_depth_at_low_tide_m}
\usage{
ss_get_colour_palette(dat)
}
\arguments{
\item{dat}{Data frame with at least one column. The column name must include
the string "low_tide".}
}
\value{
Returns a vector of hex colours from the viridis palette (Option D,
  direction = -1).
}
\description{
Returns colour palette based on the unique values of
sensor_depth_at_low_tide_m
}
\details{
Returns a discrete colour scale palette from the viridis package
  (Option D). If there are 6 or less unique values of
  \code{sensor_depth_at_low_tide_m}, the palette will have 6 colours. If
  there are more than 6 unique values of \code{sensor_depth_at_low_tide_m},
  there will be one colour for each depth.
}
\seealso{
Other plot:
\code{\link[=ss_plot_variables]{ss_plot_variables()}}
}
\author{
Danielle Dempsey
}
\concept{plot}
