% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_ggplot_variables.R
\name{ss_ggplot_variables}
\alias{ss_ggplot_variables}
\title{Plot variables at depth}
\usage{
ss_ggplot_variables(
  dat,
  superchill = NULL,
  color_palette = NULL,
  color_col = "sensor_depth_at_low_tide_m",
  legend_name = "Depth (m)",
  legend_position = "right",
  date_breaks_major = NULL,
  date_breaks_minor = NULL,
  date_labels_format = "\%Y-\%m-\%d",
  yaxis_newline = FALSE,
  point_size = 0.25,
  convert_sn_to_factor = TRUE
)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide or long format.}

\item{superchill}{Logical argument indicating whether to adding shading to
indicate temperatures below the superchill threshold (<= - 0.7 degrees C).
If \code{superchill = NULL} (the default), shading will be applied if any
temperature values are less than or equal to the threshold.}

\item{color_palette}{Optional vector of hex colors onto which depth will be
mapped.}

\item{color_col}{Character string indicating the column to use to colour the
observations.}

\item{legend_name}{Name for the depth legend. Default is \code{legend_name =
"Depth (m)"}.}

\item{legend_position}{Position for the depth legend. Default is
\code{legend.position = "right"}.}

\item{date_breaks_major}{Intervals for major breaks. Default is selected by
\code{get_xaxis_breaks()}.}

\item{date_breaks_minor}{Intervals for minor breaks. Default selected by
\code{get_xaxis_breaks()}.}

\item{date_labels_format}{Format for the date labels. Default is YYYY-mm-dd.}

\item{yaxis_newline}{Logical argument indicating whether the units in the
y-axis label should be on a new line. Default is \code{TRUE}.}

\item{point_size}{Numeric value indicating size of points.}

\item{convert_sn_to_factor}{Logical argument indicating whether to convert
column \code{sensor_serial_number} to a factor. Only used when
\code{color_col = "sensor_serial_number"}. Set to \code{FALSE} if making
separate plots with some overlapping serial numbers. In this case, the
\code{sensor_serial_number} should be converted to a factor before calling
the plot function.}
}
\value{
Returns a ggplot object of ocean variables plotted over time and
  coloured by sensor depth.
}
\description{
Plot variables at depth
}
