% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-misc.R
\name{ss_import_path}
\alias{ss_import_path}
\title{Generate file path to import raw sensor string deployment data}
\usage{
ss_import_path(station, depl_date, path = NULL)
}
\arguments{
\item{station}{Character string of the station name. Will be converted to
lower case, and a space will be replaced with an underscore.}

\item{depl_date}{Character string of the deployment data in the order
yyyy-mm-dd.}

\item{path}{Character string of the file path to the station_folders folder.}
}
\value{
The file path for importing raw deployment data.
}
\description{
Generate file path to import raw sensor string deployment data
}
