% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_map_stations.R
\name{ss_map_stations}
\alias{ss_map_stations}
\title{Interactive map of station location(s)}
\usage{
ss_map_stations(dat)
}
\arguments{
\item{dat}{Data frame with columns \code{longitude}, \code{latitude},
\code{station}.}
}
\value{
A leaflet object.
}
\description{
Interactive map of station location(s)
}
