% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_log.R
\name{ss_parse_log}
\alias{ss_parse_log}
\title{Extract information from deployment log}
\usage{
ss_parse_log(
  log,
  deployment_dates = TRUE,
  area_info = TRUE,
  sn_table = TRUE,
  config = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{log}{Log data frame, e.g., as exported from \code{ss_read_log()}.}

\item{deployment_dates}{Logical argument indicating whether to parse and
return the deployment and retrieval dates.}

\item{area_info}{Logical argument indicating whether to parse and return the
location details.}

\item{sn_table}{Logical argument indicating whether to parse and return the
serial numbers and sensor depth.}

\item{config}{Logical argument indicating whether to parse and return the
string configuration.}

\item{verbose}{Logical argument indicating whether to print messages when
values are missing. Warnings will still be printed.}
}
\value{
Returns a list with up to 4 elements. \code{deployment_dates} is a
  data frame with two columns: \code{start_date} and \code{end_date}.
  \code{area_info} is a data frame with six columns: \code{region} or
  \code{county}, \code{waterbody}, \code{latitude}, \code{longitude},
  \code{station}, and \code{lease}. \code{sn_table} is a data frame with
  three columns: \code{log_sensor} (sensor name as recorded in the log),
  \code{sensor_serial_number}, and \code{depth}. \code{string_configuration}
  is a character string indicating how the sensor string was moored.
}
\description{
The function will attempt to parse the following columns: \code{county},
\code{waterbody},\code{station}, \code{lease}, \code{deployment_date},
\code{retrieval_date}, \code{deployment_latitude},
\code{deployment_longitude}, \code{sensor_type}, \code{sensor_serial_number},
\code{sensor_depth_m}. Missing information will be filled in with \code{NA}.
}
\details{
If \code{verbose = TRUE}, a message will be printed if there is more than one
unique entry in \code{waterbody}, \code{station}, \code{deployment_date},
\code{retrieval_date}, \code{deployment_latitude}, or
\code{deployment_longitude}.

Entries in the \code{sensor_type} column must include the string
"aquameasure", "hobo", "tidbit", or "vr2ar" (not case sensitive).

A Warning message is printed when the function does not recognize a sensor in
the log.

Code will stop with an error if \code{deployment_longitude} is a positive
value.
}
\examples{
path <- system.file("extdata", package = "sensorstrings")

log_raw <- ss_read_log(path, parse = FALSE)

log <- ss_parse_log(log_raw)

log$deployment_dates
log$area_info
log$sn_table
log$string_configuration

# only return the serial number table
ss_parse_log(log_raw, deployment_dates = FALSE, area_info = FALSE, config = FALSE)
}
