% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_log_simple.R
\name{ss_parse_log}
\alias{ss_parse_log}
\title{Extract information from deployment log}
\usage{
ss_parse_log(
  log,
  deployment_dates = TRUE,
  area_info = TRUE,
  sn_table = TRUE,
  config = TRUE
)
}
\arguments{
\item{log}{Log data frame, e.g., as exported from \code{ss_read_log_simple()}.}

\item{deployment_dates}{Logical argument indicating whether to parse and
return the deployment and retrieval dates.}

\item{area_info}{Logical argument indicating whether to parse and return the
location details.}

\item{sn_table}{Logical argument indicating whether to parse and return the
serial numbers and sensor depth.}

\item{config}{Logical argument indicating whether to parse and return the
string configuration.}
}
\value{
Returns a list with up to 4 elements. \code{deployment_dates} is a
 data frame with two columns: \code{start_date} (the date of deployment) and
 \code{end_date} (date of retrieval). \code{area_info} is a data frame with
 five columns: \code{county}, \code{waterbody}, \code{latitude},
 \code{longitude}, \code{station}, and \code{lease}. \code{sn_table} is a
 data frame with three columns: \code{log_sensor} (sensor name as recorded in
 the log), \code{sensor_serial_number}, and \code{depth} (sensor depth below
 the surface at low tide, from the Sensor_Depth column).
 \code{string_configuration} is a character string indicating how the sensor
 string was moored (i.e., whether the sensors float with the tide).
}
\description{
log must include the following columns (or the information will be recorded as
NA):
}
\details{
\code{county}: county where string was deployed

\code{waterbody}: waterbody where string was deployed

\code{station}: the station name

\code{lease}: the aquaculture lease number, if applicable (NA otherwise)

\code{deployment_date}: deployment date, in the order "YYYY-MM-DD"

\code{retrieval_date}: retrieval date, in the order "YYYY-MM-DD"

\code{deployment_latitude}: The latitude at which the string was deployed

\code{deployment_longitude} The longitude at which the string was deployed
(must be a negative value)

\code{sensor_type} The type of sensor

\code{sensor_serial_number} The sensor serial number

\code{sensor_depth_m}: Depth at which the sensor was deployed, in m

All other columns will be ignored.

Entries in the \code{sensor_type} column must include the string
"aquameasure", "hobo", "tidbit", or "vr2ar" (not case sensitive).

The function will stop with an Error if there is more than one eligible file
(csv, .xlsx or .xls) in the Log folder.

A warning is printed if \code{deployment_longitude} is a positive value.

A Warning message is printed when the function does not recognize a sensor in
the log.

A message is printed to the console if there is more than one unique entry in
\code{waterbody}, \code{station}, \code{deployment_date},
\code{retrieval_date}, \code{deployment_latitude}, or
\code{deployment_longitude}.
}
