% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_pivot.R
\name{ss_pivot_longer}
\alias{ss_pivot_longer}
\title{Pivot sensor data from long to wide format}
\usage{
ss_pivot_longer(dat_wide)
}
\arguments{
\item{dat_wide}{Data frame of sensor string data, as returned by
\code{ss_compile_**()} functions.}
}
\value{
Returns \code{dat} in long format. Variables (e.g., temperature,
  dissolved, salinity, and depth measured by sensor) are in a column named
  \code{variable} and the associated measurement in a column named
  \code{value}.
}
\description{
Pivot sensor data from long to wide format
}
