% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_pivot.R
\name{ss_pivot_wider}
\alias{ss_pivot_wider}
\title{Pivot sensor string data from wide to long format}
\usage{
ss_pivot_wider(dat_long)
}
\arguments{
\item{dat_long}{Data frame of sensor string data in long format, as returned
by \code{ss_pivot_longer()}.}
}
\value{
Returns \code{dat_long} in wide format, with a separate column for
  each \code{variable}.
}
\description{
Pivot sensor string data from wide to long format
}
