% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_plot_variables.R
\name{ss_plot_variables}
\alias{ss_plot_variables}
\title{Exports figure of variable(s) at depth over time.}
\usage{
ss_plot_variables(
  dat,
  colour_palette = NULL,
  date_breaks_major = NULL,
  date_breaks_minor = NULL,
  date_labels_format = "\%Y-\%m-\%d",
  standard_do_ylims = TRUE,
  standard_ph_ylims = TRUE,
  standard_sal_ylims = TRUE,
  yaxis_newline = FALSE,
  alpha = 1,
  legend_name = "Depth (m)",
  legend_position = "right"
)
}
\arguments{
\item{dat}{Water Quality data in long or wide format.}

\item{colour_palette}{Colour palette of hex colours onto which
\code{sensor_depth_at_low_tide_m} will be mapped. Default is
\code{rev(viridis(6, option = "D"))}.}

\item{date_breaks_major}{Intervals for major breaks. Default is selected by
\code{get_xaxis_breaks()}.}

\item{date_breaks_minor}{Intervals for minor breaks. Default selected by
\code{get_xaxis_breaks()}.}

\item{date_labels_format}{Format for the date labels. Default is YYYY-mm-dd.}

\item{standard_do_ylims}{If \code{TRUE}, the y-limits for dissolved oxygen
are set to c(60, 160) \% or c(0, 15) mg/L. If \code{FALSE}, the y-limits
are set to the \code{ggplot} default. Alternatively, a vector giving custom
y-limits for the dissolved oxygen panel.}

\item{standard_ph_ylims}{If \code{TRUE}, the y-limits for pH are set to
c(7.5, 8.5) pH units. If \code{FALSE}, the y-limits are set to the
\code{ggplot} default. Alternatively, a vector giving custom y-limits for
the pH panel.}

\item{standard_sal_ylims}{If \code{TRUE}, the y-limits for salinity are set
to c(25, 34) PSU. If \code{FALSE}, the y-limits are set to the
\code{ggplot} default. Alternatively, a vector giving custom y-limits for
the salinity panel.}

\item{yaxis_newline}{Logical argument indicating whether the units in the
y-axis label should be on a new line. Default is \code{FALSE}.}

\item{alpha}{Value indicating the transparency of the points. 0 is most
transparent; 1 is opaque.}

\item{legend_name}{Name for the legend. Must be a character string. Default
is \code{legend_name = "Depth (m)"}.}

\item{legend_position}{Position for the legend. Passed to \code{patchwork}.
Default is \code{legend_position = "right"}.}
}
\value{
Returns a patchwork object, a single figure with the plots for each
  variable in \code{dat} stacked in a column.
}
\description{
Output is not compatible with ggplot2 functions or ggplotly(). Use
\code{ss_ggplot_variables()} to work with those functions.
}
\seealso{
Other plot:
\code{\link[=ss_get_colour_palette]{ss_get_colour_palette()}}
}
\author{
Danielle Dempsey
}
\concept{plot}
