% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_aquameasure_data.R
\name{ss_read_aquameasure_data}
\alias{ss_read_aquameasure_data}
\title{Import data from aquaMeasure sensors}
\usage{
ss_read_aquameasure_data(path, file_name, force_POSIXct = TRUE)
}
\arguments{
\item{path}{File path to the aquaMeasure folder.}

\item{file_name}{Name of the file to import, including file extension.}

\item{force_POSIXct}{Logical argument indicating whether to convert the
timestamp column to POSIXct. Timestamp messages "undefined" and "xs after
startup (time not set)" will be coerced to \code{NA} without a warning.}
}
\value{
Returns a tibble of aquaMeasure data, with the same columns as in the
  original file.
}
\description{
Import data from aquaMeasure sensors
}
\details{
The aquaMeasure data must be saved in a folder named aquaMeasure in
  .csv or .xlsx format.

  The same data imported from .csv and .xlsx may have a string mismatch in
  the Timestamp(UTC) Component. This is because read_excel includes the
  seconds, but fread does not
}
\author{
Danielle Dempsey
}
