% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_aquameasure.R
\name{ss_read_aquameasure_data}
\alias{ss_read_aquameasure_data}
\title{Import data from aquaMeasure sensors}
\usage{
ss_read_aquameasure_data(path, file_name)
}
\arguments{
\item{path}{File path to the aquameasure folder.}

\item{file_name}{Name of the file to import, including file extension.}
}
\value{
Returns a data.frame of aquaMeasure data, with the same columns as in
  the original file.
}
\description{
Import data from aquaMeasure sensors
}
\details{
The aquaMeasure data must be saved in csv format.
}
\author{
Danielle Dempsey
}
