% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_hobo_data.R
\name{ss_read_hobo_data}
\alias{ss_read_hobo_data}
\title{Import data from HOBO and TidbiT sensors}
\usage{
ss_read_hobo_data(path, file_name)
}
\arguments{
\item{path}{File path to the HOBO folder.}

\item{file_name}{Name of the file to import, including file extension.}
}
\value{
Returns a tibble of HOBO data, with the same columns as in the
  original file.
}
\description{
Import data from HOBO and TidbiT sensors
}
\details{
The HOBO and TidBiT data must be saved in a folder named Hobo in
  .csv or .xlsx format.

  The same data imported from .csv and .xlsx will have a string mismatch in
  the Names attribute. The degree symbol in the temperature column is
  replaced by "A" when read from a .csv file.
}
\author{
Danielle Dempsey
}
