% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_hobo.R
\name{ss_read_hobo_data}
\alias{ss_read_hobo_data}
\title{Import data from hobo or tidbit csv file}
\usage{
ss_read_hobo_data(path, file_name)
}
\arguments{
\item{path}{File path to the hobo or tidbit file.}

\item{file_name}{Name of the file to import, including file extension.}
}
\value{
Returns a data frame of hobo or tidbit data, with the same columns as
  in the original file.
}
\description{
Import data from hobo or tidbit csv file
}
\details{
The data must be saved in csv format.
}
\author{
Danielle Dempsey
}
