% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_hobo.R
\name{ss_read_hobo_data}
\alias{ss_read_hobo_data}
\title{Import data from HOBO and TidbiT sensors}
\usage{
ss_read_hobo_data(path, file_name)
}
\arguments{
\item{path}{File path to the hobo file.}

\item{file_name}{Name of the file to import, including file extension.}
}
\value{
Returns a tibble of HOBO data, with the same columns as in the
  original file.
}
\description{
Import data from HOBO and TidbiT sensors
}
\author{
Danielle Dempsey
}
