% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_log.R
\name{ss_read_log}
\alias{ss_read_log}
\title{Read in deployment log}
\usage{
ss_read_log(path, parse = TRUE, verbose = TRUE)
}
\arguments{
\item{path}{File path to the log folder, or full path to the log file,
include file name and extension. The function will stop with an Error if
there is more than one eligible file (csv, .xlsx or .xls) in the log
folder.}

\item{parse}{Logical argument indicating whether to parse log into a list
used by \code{ss_compile_*} functions.}

\item{verbose}{Logical argument indicating whether to print messages when
values are missing. Warnings will still be printed.}
}
\value{
Returns a a data frame of the deployment metadata. Option to parse
  the information into a list using \code{ss_parse_log()}.
}
\description{
The log must be saved in .csv, .xlsx or .xls format. Value checks are applied
in \code{ss_parse_log()}.
}
\details{
"Old" log column names will be converted to the new standard (.e.,g
"Location_Description" will be replaced with "station").
}
