% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_log.R
\name{ss_read_log}
\alias{ss_read_log}
\title{Extract information from deployment log}
\usage{
ss_read_log(path)
}
\arguments{
\item{path}{File path to the Log folder.}
}
\value{
Returns a list with 5 elements. \code{deployment.dates} is a
  dataframe with two columns: \code{start.date} (the date of deployment) and
  \code{end.date} (date of retrieval). \code{area.info} is a dataframe with
  five columns: \code{waterbody}, \code{latitude}, \code{longitude},
  \code{station}, and \code{lease}. \code{HOBO}, \code{aM}, and \code{vemco}
  are each a dataframe with two columns: \code{SENSOR} (serial number) and
  the

  corresponding \code{DEPTH} (depth of deployment in m).
}
\description{
Extract information from deployment log
}
\details{
The log must be saved in a folder called Log in .csv, .xlsx or .xls
  format, and must include the following columns:

  \code{Deployment_Waterbody}: waterbody where string was deployed

  \code{Location_Description}: the station name

  \code{Lease#}: If located on an aquaculture site, the lease number (NA
  otherwise)

  \code{Deployment}: The deployment date, in the order "Ymd"

  \code{Retrieval}: The retrieval date, in the order "Ymd"

  \code{Logger_Latitude}: The latitude at which the string was deployed

  \code{Logger_Longitude} The longitude at which the string was deployed

  \code{Logger_Model} The type of sensor; see below for options

  \code{Serial#} The sensor serial number

  \code{Sensor_Depth}: Depth at which the sensor was deployed

  All other columns will be ignored.

  Entries in the \code{Logger_Model} column can be "HOBO Pro V2", "HOBO DO",
  "TidbiT MX2303", "TidbiT MX2203"  "aquaMeasure DOT", "aquaMeasure SAL",
  "aquaMeasure SST", or "VR2AR". (Some mis-spellings are accepted: "HOBO pro
  V2", "HOBO_Pro_V2", "aquameasure DOT", "aquameasure SAL", "aquameasure
  SST")

  A Warning message is printed to the console when the function does not
  recognize a sensor in the log.

  A message is printed to the console when Hobo, aquaMeasure, or Vemco
  sensors are not found in the log.

  A message will be printed to the console if there is more than one unique
  entry in \code{Deployment_Waterbody}, \code{Location_Description},
  \code{Deployment}, \code{Retrieval}, \code{Logger_Latitude}, or
  \code{Logger_Longitude}.

  If there is more than one eligible file (csv, .xlsx or .xls) in the Log
  folder, the function will stop with an error.
}
\seealso{
Other compile: 
\code{\link{ss_compile_aquameasure_data}()},
\code{\link{ss_compile_deployment_data}()},
\code{\link{ss_compile_hobo_data}()},
\code{\link{ss_compile_vemco_data}()}
}
\author{
Danielle Dempsey
}
\concept{compile}
