% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_nsdfa_metadata.R
\name{ss_read_nsdfa_metadata}
\alias{ss_read_nsdfa_metadata}
\title{Import TempMetaData tab from the NSDFA tracking sheet}
\usage{
ss_read_nsdfa_metadata(path)
}
\arguments{
\item{path}{Path to the NSDFA tracking sheet (include file name and
extension).}
}
\value{
Returns data frame of NSDFA tracking sheet sensor string metadata.
}
\description{
Import TempMetaData tab from the NSDFA tracking sheet
}
\details{
Reads in the TempMetaData tab from the NSDFA tracking sheet and
  corrects known errors (e.g., standardizes station and waterbody spellings,
  fixes deployment dates, etc.).

  **Might want to add 2021-08-27 1042 deployment
}
