% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_vdat.R
\name{ss_read_vdat_data}
\alias{ss_read_vdat_data}
\title{Import data from VR2AR csv files exported from Fathom}
\usage{
ss_read_vdat_data(path, file_name = NULL)
}
\arguments{
\item{path}{File path to the vdat folder, or full path to the data file
including file name and extension.}

\item{file_name}{Name of the file to import, including file extension. Should
be \code{NULL} if the file name is included in the \code{path} argument.}
}
\value{
Returns a data frame of vr2ar data, with the same columns as in the
  original file.
}
\description{
Import data from VR2AR csv files exported from Fathom
}
\details{
For VR2AR data offloaded in the new .vdat format and exported with
  Fathom software. The data must be saved in csv format.
}
\examples{
path <- system.file("testdata", "vdat", package = "sensorstrings")

# folder and file name
vdat_raw <- ss_read_vdat_data(path, "VR2AR-69_555438.csv")
head(vdat_raw)

# or the full path to the file
vdat_raw <- ss_read_vdat_data(file.path(path, "VR2AR-69_555438.csv"))
}
