% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_write_report_table.R
\name{ss_write_report_table}
\alias{ss_write_report_table}
\title{Writes deployment table for county report}
\usage{
ss_write_report_table(dat, keep_waterbody = FALSE)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format.}

\item{keep_waterbody}{Logical value indicating whether to keep the
\code{Waterbody} column.}
}
\description{
Writes deployment table for county report
}
