% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-plot.R
\name{ss_xaxis_breaks}
\alias{ss_xaxis_breaks}
\title{Returns nice major and minor breaks and label format based on timespan of the
data}
\usage{
ss_xaxis_breaks(dat)
}
\arguments{
\item{dat}{Data frame with at least one column: \code{timestamp_utc}
(POSIXct).}
}
\value{
Returns a dataframe with 1 observation of 3 variables
  \code{date_breaks_major}, \code{date_breaks_minor},
  \code{date_labels_format}.
}
\description{
Returns nice major and minor breaks and label format based on timespan of the
data
}
